+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Updated several dev dependencies
+- App_id and repository name changed to cfg_share_links (previously cfg-share-links)
 
 ## [1.0.1] - 2022-02-14